feat(budget-management): 新增预算科目功能并整合 MyBatis-Plus
- 新增 BudgetItem相关的实体类、Mapper、Service 及 Controller
-集成 MyBatis-Plus并配置分页插件
- 更新数据库操作逻辑，支持预算科目状态批量修改
- 优化查询结果，增加预算科目类型和分组名称字段
- 移除冗余的 MyBatis 配置

+++ src/main/java/org/psd/budget_management/controller/BudgetItemTypeController.java
@@ -73,13 +73,13 @@ public class BudgetItemTypeController {
     }
 
     /**
-     * 删除数据
+     * 批量删除数据
      *
      * @param idList 主键结合
      * @return 删除结果
      */
     @DeleteMapping
-    public Result deleteByIds(@RequestParam("idList") List<Long> idList) {
+    public Result deleteByIds(@RequestParam("idList") List<Integer> idList) {
         return new Result(200, "执行成功", this.budgetItemTypeService.removeByIds(idList));
     }
 }
